refactor: splitting serializable trait
