+++ crates/committer/src/deserialization.rs
@@ -1,4 +1,3 @@
 pub mod cast;
-pub mod errors;
 pub mod read;
 pub mod types;
