+++ crates/committer/src/deserialization/read.rs
@@ -1,6 +1,6 @@
-use crate::deserialization::errors::DeserializationError;
 use crate::deserialization::types::Input;
 use crate::deserialization::types::RawInput;
+use crate::storage::errors::DeserializationError;
 
 #[cfg(test)]
 #[path = "read_test.rs"]
