+++ crates/committer/src/deserialization/read_test.rs
@@ -1,13 +1,13 @@
 use pretty_assertions::assert_eq;
 use std::collections::HashMap;
 
-use crate::deserialization::errors::DeserializationError;
 use crate::deserialization::types::{
     ContractAddress, ContractState, Input, StarknetStorageKey, StarknetStorageValue, StateDiff,
 };
 use crate::hash::hash_trait::HashOutput;
 use crate::patricia_merkle_tree::filled_tree::node::{ClassHash, CompiledClassHash, Nonce};
 use crate::patricia_merkle_tree::types::TreeHeight;
+use crate::storage::errors::DeserializationError;
 use crate::storage::storage_trait::{StorageKey, StorageValue};
 use crate::types::Felt;
