+++ crates/committer/src/patricia_merkle_tree/errors.rs
@@ -2,7 +2,7 @@ use thiserror::Error;
 
 use crate::patricia_merkle_tree::node_data::leaf::LeafDataTrait;
 use crate::patricia_merkle_tree::types::NodeIndex;
-use crate::storage::errors::{SerializationError, StorageError};
+use crate::storage::errors::{DeserializationError, StorageError};
 
 use crate::patricia_merkle_tree::filled_tree::node::FilledNode;
 
@@ -12,7 +12,7 @@ pub(crate) enum OriginalSkeletonTreeError {
     #[error(
         "Failed to deserialize the storage value: {0:?} while building the original skeleton tree."
     )]
-    Deserialization(#[from] SerializationError),
+    Deserialization(#[from] DeserializationError),
     #[error(
         "Unable to read from storage the storage key: {0:?} while building the \
          original skeleton tree."
